requirements.txt
pyyaml
openai
python-dotenv
requests
